+++ src/providers/formlyApiCheck.test.js
@@ -3,23 +3,24 @@ module.exports = ngModule => {
   describe('formlyApiCheck', () => {
     beforeEach(window.module(ngModule.name));
 
-    describe('formlyFIeldOptions', () => {
-      let formlyFIeldOptions;
+    describe('formlyFieldOptions', () => {
+      let formlyFieldOptions;
       beforeEach(inject((formlyApiCheck) => {
-        formlyFIeldOptions = formlyApiCheck.formlyFieldOptions;
+        formlyFieldOptions = formlyApiCheck.formlyFieldOptions;
       }));
 
-      it(`should pass when validation.messages is an object of functions`, () => {
+      it(`should pass when validation.messages is an object of functions or strings`, () => {
         const options = {
           key: '♪┏(・o･)┛♪┗ ( ･o･) ┓♪',
           validation: {
             messages: {
               thing1() {
-              }
+              },
+              thing2: '"Formly Expression"'
             }
           }
         };
-        const result = formlyFIeldOptions(options);
+        const result = formlyFieldOptions(options);
         expect(result).to.be.undefined;
       });
     });